release 3.1.179 source code

+++ services/kafka/src/main/java/com/huaweicloud/sdk/kafka/v2/model/ShowInstanceTopicDetailRespPartitions.java
@@ -26,17 +26,17 @@ public class ShowInstanceTopicDetailRespPartitions {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "leo")
 
-    private Integer leo;
+    private Long leo;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "hw")
 
-    private Integer hw;
+    private Long hw;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "lso")
 
-    private Integer lso;
+    private Long lso;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "last_update_timestamp")
@@ -82,7 +82,7 @@ public class ShowInstanceTopicDetailRespPartitions {
         this.leader = leader;
     }
 
-    public ShowInstanceTopicDetailRespPartitions withLeo(Integer leo) {
+    public ShowInstanceTopicDetailRespPartitions withLeo(Long leo) {
         this.leo = leo;
         return this;
     }
@@ -91,15 +91,15 @@ public class ShowInstanceTopicDetailRespPartitions {
      * **参数解释**： 分区leader副本的LEO（Log End Offset）。 **取值范围**： 不涉及
      * @return leo
      */
-    public Integer getLeo() {
+    public Long getLeo() {
         return leo;
     }
 
-    public void setLeo(Integer leo) {
+    public void setLeo(Long leo) {
         this.leo = leo;
     }
 
-    public ShowInstanceTopicDetailRespPartitions withHw(Integer hw) {
+    public ShowInstanceTopicDetailRespPartitions withHw(Long hw) {
         this.hw = hw;
         return this;
     }
@@ -108,15 +108,15 @@ public class ShowInstanceTopicDetailRespPartitions {
      * **参数解释**： 分区高水位（HW，High Watermark）。 **取值范围**： 不涉及
      * @return hw
      */
-    public Integer getHw() {
+    public Long getHw() {
         return hw;
     }
 
-    public void setHw(Integer hw) {
+    public void setHw(Long hw) {
         this.hw = hw;
     }
 
-    public ShowInstanceTopicDetailRespPartitions withLso(Integer lso) {
+    public ShowInstanceTopicDetailRespPartitions withLso(Long lso) {
         this.lso = lso;
         return this;
     }
@@ -125,11 +125,11 @@ public class ShowInstanceTopicDetailRespPartitions {
      * **参数解释**： 分区leader副本的LSO（Log Start Offset）。 **取值范围**： 不涉及
      * @return lso
      */
-    public Integer getLso() {
+    public Long getLso() {
         return lso;
     }
 
-    public void setLso(Integer lso) {
+    public void setLso(Long lso) {
         this.lso = lso;
     }
